Move swap-domain helpers to appropriate folder
