+++ src/cow-react/swap/hooks/useSwapButtonContext.ts
@@ -16,7 +16,7 @@ import { useApproveCallbackFromTrade } from 'hooks/useApproveCallback'
 import { useGnosisSafeInfo } from 'hooks/useGnosisSafeInfo'
 import { useIsSwapUnsupported } from 'hooks/useIsSwapUnsupported'
 import { ApproveButtonProps } from 'pages/Swap/components/ApproveButton'
-import { getSwapButtonState } from 'pages/Swap/helpers/getSwapButtonState'
+import { getSwapButtonState } from 'cow-react/swap/helpers/getSwapButtonState'
 import { SwapButtonContext } from 'pages/Swap/components/SwapButton/SwapButton'
 import { useGetQuoteAndStatus } from 'state/price/hooks'
 import { OperationType } from 'components/TransactionConfirmationModal'
